refactor: 열거형 클래스명 type으로 통일 및 패키지로 묶기
