+++ src/main/java/be/podor/review/dto/ReviewRequestDto.java
@@ -1,7 +1,7 @@
 package be.podor.review.dto;
 
 import be.podor.review.model.reviewInfo.ScoreEnum;
-import be.podor.theater.model.FloorEnum;
+import be.podor.theater.model.type.FloorType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ReviewRequestDto {
     private Long theaterId;
-    private FloorEnum floor;
+    private FloorType floor;
     private String section;
     private String row;
     private Integer seat;
