+++ src/main/java/be/podor/theater/model/type/FloorType.java
@@ -1,4 +1,4 @@
-package be.podor.theater.model;
+package be.podor.theater.model.type;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum FloorEnum {
+public enum FloorType {
     FIRST("1F", 1),
     SECOND("2F", 2),
     THIRD("3F", 3);
@@ -15,10 +15,10 @@ public enum FloorEnum {
     private final int num;
 
     @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
-    public static FloorEnum from(String floor) {
-        for (FloorEnum floorEnum : FloorEnum.values()) {
-            if (floorEnum.getFloor().equals(floor)) {
-                return floorEnum;
+    public static FloorType from(String floor) {
+        for (FloorType floorType : FloorType.values()) {
+            if (floorType.getFloor().equals(floor)) {
+                return floorType;
             }
         }
         return null;
