+++ src/main/java/be/podor/theater/model/type/GradeType.java
@@ -1,4 +1,4 @@
-package be.podor.theater.model;
+package be.podor.theater.model.type;
 
 public enum GradeType {
     VIP, R, S, A;
